sla: Update balancers and add memory based load balancing
This patch cleans up the load balancing modules and creates
a new base class for them - CpuAndMemoryBalancingPolicyUnit.
It also cleans up the naming conventions in the code so we
are no longer confused by "overLoadedCPU" method getting
data about idle hosts.

The base class supports primary and secondary reason for
balancing, but does not force the content. All inherited
classes then provide their own lists for primary and
secondary source and destination hosts.

The base class provides parametrized methods for getting
CPU and memory over-, under- and normally loaded hosts.
These methods are meant to be called from the derived classes
when preparing the primary and secondary lists.

Migration is never attempted to CPU or memory overloaded
hosts (not even in power saving).

The Even Distribution unit is then updated to use Memory
as the secondary criteria when no CPU based load balancing
is needed (or possible).

The Power saving unit is updated in the same way, except
it tries to migrate VMs from over and under utilized hosts
to the middle group - normally utilized hosts. This should
behave the same as the old code with the exception of having
the memory information used as well.

Change-Id: I1fe13267feca89ab6c8fb9d85656f05930d0b333
Signed-off-by: Martin Sivak <msivak@redhat.com>