restapi: Fix XML schema for fencing agents
The "Agents" XML complex type corresponds to a collection, not to an
entity. But it is defined as an extension of the "BaseResource" type,
which corresponds to entities, not to collections. This breaks our rules
for XML schema and induces errors in the Java SDK, as the generated type
"Agents" is not within the bounds of the type parameters used by the
generic type "CollectionDecorator". This patch fixes the "Agents" type
so that it extends "BaseResources".

Change-Id: Ia5eed94f34ce7fafd98c362356e333db117668be
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>