webadmin: Create VM from Templates list view
Adding a new UI action "New VM" in the Templates list view.
It is enable only when a single template is selected.
When the user clicks on this action, the "New VM" dialog appears.
In the template drop down only the selected template from the list and
all it versions will be available. The selected item in the drop down
will be the template highlighted in the list view.

Change-Id: I0e5fd7e84bbea9405bf7d9e087b4281866daa954
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1117484
Signed-off-by: Fred Rolland <frolland@redhat.com>