restapi: Add action to refresh host capabilities
Currently the only way to refresh the host capabilities is to add the
"force" matrix parameter to the GET request of the host. This is mixing
resource retrevial and actions, which isn't good, and isn't supported by
the Python SDK. This patch adds a new "refreshcapabilities" action. The
existing "force" parameter is preserved for backwards compatibility, but
will be likely removed in the future.

Change-Id: I4ce36ac7f615b3c1b8a5b460ee1c1a482d6240fa
Bug-Url: https://bugzilla.redhat.com/1197788
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>