restapi: Allow creation of new disks as copy of any existing disk
This patch allows to copy a VM, Template or floating Disk to a different or
same storage domain.

A new optional argument has been added: "name". If the name is set, the copied
disk's alias will be set to the given value. If not, the alias of the copied
disk will be used.

Example:
POST to .../api/disks/<disk id>/copy

<action>
<storage_domain id="14e6e9f3-9fe9-493b-b9ba-793cb441f9ad" />
<disk><alias>newDiskAlias</alias></disk>
</action>

Bug-Url: https://bugzilla.redhat.com/1132066
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>
Change-Id: I8931d2464d6ad17a80d9b394f16d01a2c47d7575