webadmin: cow-sparse base volumes of templates
Allow the user to specify the format of template volumes.
The type of the volumes is generated according to its format and the
storage type according to the following rules:
QCow -> SPARSE
RAW + File -> SPARSE
RAW + Block -> PREALLOCATED
That way, the user can set base volume of a template to be QCOW+SPARSE
which was not possible to set before.

Change-Id: Id269321c793e4b2af2c8efb153c37a3721f8cbf7
Bug-Url: https://bugzilla.redhat.com/1168576
Signed-off-by: Arik Hadas <ahadas@redhat.com>