getVMList: introducing onlyUUID parameter
Due to vdsm sending only an array of vmids in a response to getVMList
the engine asks separately using VM.getStats which creates unnecessary
load. We need to make sure that for newer engines vdsm sends map of
vmids and status and for older engines an array as it was before. In
order to achieve it we introduce new parameter 'onlyUUID' for getVMList
which is ignored by older vdsms and it makes sure that newer vdsms send
a map which fixes performance issue.

Change-Id: I51dea68e89a2c1685d16ba1921e1373eeb721860
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Bug-Url: https://bugzilla.redhat.com/1198248
Related-To: https://gerrit.ovirt.org/#/c/38449