engine: Updating the maintenance_reason on MaintenanceNumberOfVdssCommand
SetVdsStatusVdsCommand gets the maintenanceReason as a parameter and stores
it in the db. If the reason is not set, a null will be stored.

When the maintenance command is executed the reason (if there is one) is
passed to SetVdsStatusVdsCommand.
Since all the vds status changes are done via SetVdsStatusVdsCommand, when
the status of the vds is changed to another status the maintenance reason
will be nullified (since other places that executes SetVdsStatusVdsCommand
doesn't pass maintenanceReason via the parameters).

Change-Id: If1ab2e0b3e2cc154d14282bc38b158b0b4d195fe
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>