userportal, webadmin: switch to GwtBootstrap3CDN module
(Patch 1 of 12)

Switched to GwtBootstrap3CDN module from which to inherit gwt-common,
userportal, and and webadmin. Previously used GwtBootstrap3NoTheme.

The effect of this change is to stop gwtbootstrap3 from injecting jquery.js
and bootstrap.js. As we move to a move modular UI, we have decided to control
JavaScript and CSS assets outside of GWT as much as possible. If we want to
support a newer version of jquery in the future (for, say, UI plugin usage),
we won't be tied to the version that ships with gwtbootstrap3.

Added bootstrap.min.js and jquery.min.js includes in javascripts.tag to
replace the gwtbootstrap3-injected copies of bootstrap.js and jquery.js.
Also added patternfly.min.js, which is not currently required but is useful.
It includes script for a few PatternFly widgets that we do not yet use.

Change-Id: I77f50562ea8bc67f59246eb68a17f87584a5fd9b
Signed-off-by: Greg Sheremeta <gshereme@redhat.com>