core: do not use distinct if sort expr have func
A former patch made by Omer F added distinct to the search in order to
prevent duplicates.
In the case the search is in the format :

SELECT DISTINCT * FROM T .... ORDER BY f({arguments})

There is a problem since the DISTINCT is performed first and then the
sorting, so f is unknown since it is not part of the result set.

This patch checks if the sort clause has a function call and removes the
DISTINCT if it finds one.

Also added a unit test for this case.

Change-Id: I7c036b2b9ee94266b6e3df54f2c50167e454ed6a
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1198506
Signed-off-by: emesika <emesika@redhat.com>