webadmin: DataCenterStorageListModel update availability refactor.
Currently every time a selected DC is changed, or whenever the
availability of storage domains of a DC should be updated in the
DC's storage tab, the program runs through all the domains all over
again for each boolean value it needs. (e.g. is there a master
domain, is there an ISO domain, etc...).

This patch replaces the multiple iterations into one. at the end
of the iteration all the boolean values will be set.

This is a better design for future boolean values to add in code
redundancy and clearness aspects.

Change-Id: I88abe054994257e2d3bd781d4d3b53aece75cd2c
Signed-off-by: Amit Aviram <aaviram@redhat.com>