restapi: Add VM parameters to VmPool collection
A majority of VM parameters that are available in the UI when adding VM Pool
were missing from REST API VmPool collection.

Added new <vm> subelement with all the VM fields into the <vmpool> element.
The values of these fields override the values of the corresponding template
parameters. The structure is:

<vm_pool>
 <size>50</size>
 ...
 <template id="xxx"/>
 <vm>
  <some_overriden_field_a>new value</some_overriden_field_a>
  <some_overriden_field_b>new value 2</some_overriden_field_b>
 </vm>
</vm_pool>

Description on oVirt Wiki:
http://www.ovirt.org/Features/Vm_Parameters_in_REST_API_for_Vm_Pools

Change-Id: I791791a99abdf258f42b3e06288567f1b6c0bbf4
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1195167
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>