frontend: Add infrastructure for Console Foreign Menu
Added necessary hooks to implement outputting the [ovirt] section
into VV file. Please note this just related to frontend infrastructure
and contains no API commitments (does not alter VV file output in any way).

Moved the RestApiSessionManager to gwt-common project so it would be
accessible from both webadmin and userportal.

Due to the fact that uicommon does not see the generated gwt-p events
and handlers we need to attach the listener for RestApiSessionAcquiredEvent
in UiCommonDefaultTypeResolver. To prevent excessive number of event
listeners attached we cache the created SpiceNativeImpl and
VncNativeImpl.

Change-Id: I0e0cc2a2fd462bd8dc9085b300ec12fe46ff4518
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=975730
Signed-off-by: Martin Betak <mbetak@redhat.com>