engine: host_device.vm_id should be engine only field
The vm id that is reported by the vdsm is the one that was passed to it by
the engine.
If a device was attached to a vm via a hook, a vm_id won't be reported.
It means, the vdsm.vm_id doesn't have any extra value and can be ignored
by the engine.

Change-Id: I6865c2414db3ff9d92b7c51f4523d85176e5b0c1
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>