core: introduce CinderBroker
Introducing CinderBroker class which handles all invocations
of Cinder sdk actions (e.g. createVolume/deleteVolume/etc).
The CinderBroker adds VdcBllErrors.CINDER_ERROR to executeFailedMessages
in case of an OpenStackResponseException.
This generic error is part of VdsmErrors as it relates to errors that
may occur in the execute phase. The messages are sent in plain string
(as opposed to codes), hence, translation is not applicable.

Change-Id: If71305412188be95dc88bc0d8646958a3d3e6aef
Bug-Url: https://bugzilla.redhat.com/1185826
Signed-off-by: Daniel Erez <derez@redhat.com>