engine: Add query for retrieving system provider
The query retrieves the provider which is associated by
the ovirt-engine server. The Foreman providers within
the system are scanned for having a content host with the
same host name as resolved for the ovirt-engine server.

Dynamic scanning will reduce the complexity of managing
external resource id for the host which represents the
ovirt-engine server and maintaining it.

Change-Id: I22bbb146a94c3f93cf5134f5f2e6d509f303c585
Bug-Url: https://bugzilla.redhat.com/1157377
Signed-off-by: Moti Asayag <masayag@redhat.com>