core: introduce DiskContentType
Generalize the ovf-store into a content-type of disk. Ovf-store is just
one example for a case where decisions are taken according to the type
of the content that is stored inside a disk. Another example would be
memory dump. So this patch replaces the ovf-store specific fields with
an enum that reflects the type of the content inside a disk.

Change-Id: I410a346cf0be32a99e10e503c26eb29b3688e748
Signed-off-by: Arik Hadas <ahadas@redhat.com>