core: Requesting to consume quota while extending a disk fix.
Until now when the user wanted to extend a disk, there was no
request generated for quota consumption. This fact let the user to
extend the disk's size as much as he likes when there are quota
limitations.

This patch adds a request for quota consumption when disks are
extended.

Change-Id: Ib31206b0a9172b35217845b01253951d93f380f1
Bug-Url: https://bugzilla.redhat.com/1203262
Signed-off-by: Amit Aviram <aaviram@redhat.com>