webadmin: Fix System tree when in Gluster-only mode
This patch adds Hosts and Volumes nodes needed in
Gluster-only mode to the System tree root. Previously only
System and Sessions were shown in Gluster-only mode
(because there is no datacenter in Gluster-only mode, and
datacenter's children were being filtered out of the tree).

This patch uses the following approach:
* Filter out the nodes as usual as per the associated ApplicationMode.
* If it is applicable to the current ApplicationMode
  * As a node is removed, attach its child nodes to its parent, thus
    promoting those children to peers.
  

Change-Id: I7c85bc2b78f885511a8e53212984d93343ffd6a8
Signed-off-by: Anmol Babu <anbabu@redhat.com>