core: vmsMonitoring - persist vms data in refresh vms phase
the new structure of the host-vms monitoring caused the save vms info to
db to be called later in the cycle than it used to be.

this caused issues in operations that executed in afterVMsRefreshTreatment, that
assumed the db is updated.

in this patch the save-to-db moved back into the refreshVmStats from the
afterVMsRefreshTreatment, also with processExternallyManagedVms and
processVmsWithDevicesChange that originally happened before the save to
db.

Change-Id: I5e69b0ba20588db55bb40e9ba6ca5472c3efe687
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>