core: use single thread on update sla policies
when vms monitoring cycle finish, it is calling update sla policies in a
new thread.
the new vms monitoring changed the call to be per-vm instead of one call
for all vms that succeeded to run.

in this patch the previous approach restored, calling once with list of
vms, resulting only one thread for all the vms.

Change-Id: Id7f92bf64d771e351f2aab46874515eb70a713fa
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>