restapi: Always use "general" value of VdcFenceOptionTypes
When the RESTAPI populates the "power_managers" section of the
capabilities resource it needs to know the option types used by the
power management options. Currently it does this retriving the version
specific version of the "VdcFenceOptionTypes" configuration parameter.
But the database only contains the "general" value of this parameter,
not the version specific values. This means that the RESTAPI will always
get the default value specified in the "ConfigValues" class, and this
doesn't contain some recently added types, like "encrypt_options", and
that triggers an exception. To solve this problem this patch changes the
RESTAPI so that it will always retrieve the "general" version of this
configuration parameter.

Change-Id: I140fac538ce1d7f146943ce5d8c80a2a787a5fc4
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>