core: [RFE]Backup Awareness - Config
Adding backup awareness to engine

Currently backups are offered only as best practice, there is no alert
or event that informs the user that no backup was done or that a long
time passed from the last successful backup and he should backup the
application database again. In case of an accident, this may lead to
long engine downtime and lots of efforts restoring the engine DB if no
backup was done or only an old backup exists

The feature will enable to track backups done using engine-backup
utility and to alert the administrator if a pre-configurable time had
passed from the last successful backup.

Feature page : http://www.ovirt.org/Features/BackupAwareness

This patch adds the configuration values needed for the feature:

BackupCheckPeriodInHours the period on which we check for a fresh backup
 default : 6, set to -1 to disable this feature
BackupAlertPeriodInDays the max number of days allowed without a fresh
backup.
 default : 1

Change-Id: I1ddd8c817742406d8206624599a82d8ff36b26ee
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1188127
Signed-off-by: Eli Mesika <emesika@redhat.com>