engine: Double Notification Received (#1192528)
When a Highly-Available VM goes down with an error,
the engine attempts to re-run it. If the first attempt
fails, the engine immediately attempts once more.

Subscribers which subscribed to the notification:
"VDS_INITIATED_RUN_VM_FAILED" receive notification about
both events, even though they are milliseconds apart
and one notification should be enough.

This bug was solved by setting a 'flood limit' of 5
seconds on this event.

Change-Id: Ib97d2e7f456f13331f843591716f280b9f17d1ca
Bug-Url: https://bugzilla.redhat.com/1192528
Signed-off-by: Ori Liel <oliel@redhat.com>