webadmin: fix import/general sub-tab
The general sub-tab in the import dialog was broken by recent changes in
the frontend - it didn't present the VM information.

This patch fixes the sub-tab that shows general information on VM in the
import dialog. I changed the sub-tab to be separate component that
doesn't inherit the VM general sub-tab in the virtual machines tab
because they differ in:

1. Layout - in the sub-tab which is presented in the import dialog, we
don't need to have the alerts tab since warnings/errors will be shown
differently

2. Items - the presented fields are different and over-time they are
expected to be changed in a way that they will be even more different

3. Configurability - in the import sub-tab some fields are editable.
currently only the name and the operating system but there should be
more, while in the general sub-tab nothing is editable

Change-Id: Ied6e9619ba6f539e42600d3650adff2b9272ac9e
Signed-off-by: Arik Hadas <ahadas@redhat.com>