webadmin: do not persist RunOnce boot sequence accidentally
VmDeviceCommonUtils.updateVmDevicesBootOrder() method was using RunOnce
boot sequence if vm.isRunOnce() is true. When called from
VmDeviceUtils.updateBootOrderInVmDeviceAndStoreToDB() this caused
RunOnce boot sequence to be persisted when attaching/detaching disks,
adding/removing NICs etc.

Do avoid this undesired behavior,
VmDeviceCommonUtils.updateVmDevicesBootOrder() is no more using RunOnce
boot sequence. The RunOnce boot sequence is passed to it only from
VmInfoBuilder.buildVmBootSequence() that is called when VM is started.
The result is used to boot the VM and never persisted.

Note that VmDynamic.bootSequence is still persisted, this is done in
another place.

Change-Id: Ibfa685f8d81b67eff41234a450185879f5bbccfd
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1209038
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>