events: Refine Host Activation Message (#1193060)
When user activates a host, ActivateVdsCommand doesn't
perform the actual activation; it only changes the host
status. The host is activated later asynchronously.

When the command terminates it audits:
Host X activated by user Y.

This is not true, and if some error prvents the actual
activation, the audit log is confusing. So the auditing
was changed to:

Activation of host X initiated by user Y.

Bug-Url: http://bugzilla.redhat.com/1193060
Change-Id: I698dac3797c957ed3dabe33490482e334d9d27d8
Signed-off-by: Ori Liel <oliel@redhat.com>