core: Fix AddVmTemplate unreleased lock
AddVmTemplate child command caused a lock lost in some conditions,
the problem occurred because the child command got the parent command
context instead of a clone, when the child command cleared the lock
(before it started the execution) it caused the parent command to lose
that lock as well.

The clone creation was missing in CommandBase::insertAsyncTaskPlaceHolders(),
because this mechanism was removed in 3.6, and because the current problem
seems to only affect addVmTemplate, this fix is local only to AddVmTemplate
so changes for the infrastructure are minimal.

Change-Id: I3e1f0c04094866bf0bb7b36a6e166280939a2d8c
Bug-Url: https://bugzilla.redhat.com/1203307
Signed-off-by: Eldan Shachar <eshachar@redhat.com>