engine: add feature comptability check for VDS
 Some times checking for the cluster compatibility version is not
enough to support some features across multiple comptability
versions. So we need to have some machanism while adding a node
to cluster to check that vdsm supports all the required features
required for the cluster.

 This patch adds feature compatability check for VDS. List of all
additional features will be maintained in a master table. List of
features will be shown in the new/edit cluster dialog based on the
compatability version selected as well as the services gluster/virt
enabled.  User can select the features he wants in the cluster and
cluster will be initialized with that set of feature during
cluster creation. When a node is added to the cluster, it checks
that all the required features for the cluster are supported by vdsm.
This is done in a generic way so that it can used in different
features in future.

Currently it is used to check whether the vdsm supports gluster37
features. gluster37 features are supported in 3.6 cluster by default.
It is also supported in 3.5 cluster with this feature check.

Change-Id: Icba02b189a169bc676e0c5f47f7aaf394f0b49a6
Signed-off-by: Ramesh Nachimuthu <rnachimu@redhat.com>