restapi: Remove empty bodies from RSDL metadata file
Currently the RSDL metadata file contains the definition of request
bodies even if they are completely empty. For example:

  body:
    parameterType: null
    signatures: []

This is currently needed because otherwise the list of signatures would
be "null" instead of an empty list, and that would induce some changes
in the generated RSDL. To avoid all these empty specifications (that
make the already giantic RSDL metadata file even larger) this patch
changes the RSDL builder so that it assigns default values to the
metadata after loading it (replacing "null" with empty lists, for
exampe) and also removes the empty specifications from the RSDL metadata
file. The generated RSDL document doesn't change.

Change-Id: I1a95ca58b83750886429bce29a0e8e9188c76d48
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>