restapi: Implicity add "async" and "grace_period.expiry" to actions
Currently the "async" and "grace_period.expiry" body parameters are
explicitly documented in the RSDL metadata file. But these parameters
are available for all actions, and missing from the documentation of
some of them. As a result the Python SDK doesn't document them
correctly, and the corresponding options aren't available in the CLI.
This patch changes the RESTAPI so that it will implicitly add these two
parameters to the RSDL for all actions.

Change-Id: I1fec7e64d0d2a33708038debd90bff17cba7b449
Bug-Url: https://bugzilla.redhat.com/1211588
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>