core: QuotaStorageDependent removal in ExtendImageSizeCommand.
ExtendImageSizeCommand currently implements QuotaStorageDependent which deals with
quota consumptions. The command is being called in UpdateVmDiskCommand
in case the disk size is extended, but never reaches the actual consume
part in the code as it is constructed as an unquota-dependent command.

Moreover, UpdateDiskCommand deals with any case of quota consumption
(introduced in https://gerrit.ovirt.org/#/c/39330/,
BZ: https://bugzilla.redhat.com/1203262 )

So dealing with that is not needed in that command.

Change-Id: I940ab5817cb09e4f6275cbda713ecc37a4bd74fc
Signed-off-by: Amit Aviram <aaviram@redhat.com>