engine, webadmin: use slave host id instead of slave host name
The canDoAction of create georep session command checks if
host is monitored by engine and tries to  get host based on the
passed host details.

Previously,this used to be obtained using host name and UI
used to pass the host name which is VDS#getName but the
backend expected dns resolvable name which is provided by
VDS#getHostName. The UI can't pass VDS#getHostName because
its not available in there. The UI lists vds as obtained from
the bricks of eligible/non-eligible volumes. The volume can
provide only brick servernames and server ids.

So, its better to obtain vds based on its id rather than on
name.

The above reasons required this change to be introduced.

Change-Id: I4440eddbc2616c7ca89feaae3ba0e67d93562213
Signed-off-by: Anmol Babu <anbabu@redhat.com>