core: redesign MemoryStorageHandler to use filters
This patch doesn't change the behavior of MemoryStorageHandler.
Instead, it changes the way we select storage domains for memory
volumes to use filters. Each filter verifies that every domain meets a
specific condition, and removes those that don't from the list of
candidates.
This way, one can simply add a new filter to the list and add new logic
to the domain selection.

Change-Id: Iff7094a2351ab0755ff21ba3ccfc4b8bc5cfff6c
Bug-Url: https://bugzilla.redhat.com/1186230
Signed-off-by: Idan Shaby <ishaby@redhat.com>