restapi: Remove redudant parameter types from RSDL metadata
The "resetemulatedmachine" operation explicitly documents the
"Content-Type" and "Correlation-Id" headers, but these are already
implicitly added to all operations. The net result is that they appear
twice in the generated RSDL document, and this induces a duplicated
parameter in the code generated for the SDKs. This patch removes the
headers from the RSDL metadata file.

Change-Id: I6c382b88300622213b20e72912614da65fdb7747
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>