gluster: Disable gluster cli based snapshot scheduling
Introduced changes to disable gluster CLI based volume snapshot
scheduling if engine is used.

It looks for the existence of a meta volume in gluster and if so,
disables the CLI based volume snapshot scheduling by setting a
required flag.

Later gluster refers the same flag and does not allow re-enabling
of CLI based volume snapshot shcheduling from command line.

Change-Id: I6bc06f246f30769c5edaf981876c2a51ddd4fffd
Bug-URL: https://bugzilla.redhat.com/1230342
Bug-URL: https://bugzilla.redhat.com/1215600
Signed-off-by: Shubhendu Tripathi <shtripat@redhat.com>