core: Save graphics devices when VM is running
Added support of saving of VM's graphics devices for the next run if
VM properties are edited when it is running. Notification to user is
displayed as well.

Added @EditableDeviceOnVmStatusField annotation to
VmManagementParametersBase.graphicsDevices field and extended the
@EditableDeviceOnVmStatusField annotation to support fields of VmDevice
and Map type in addition to Boolean. Also added missing functionality in
ProcessDownVmCommand and InstanceTypeManager to restore graphics
devices.

Change-Id: I9cce664ae9fafe2713c014a95b4569b8c7e45f8e
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1210382
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>