core: Save RNG device when VM is running
Added missing functionality to detect changes of RNG device properties
when VM is running and to save/restore them to/from NEXT_RUN snapshot.
For this purpose Optional<T> class was added to
VmManagementParametersBase. This class combines data for the device that
needs to be updated with a flag that switches the update on/off. This
change makes possible to use @EditableDeviceOnVmStatusField annotation
for rngDevice field and to combine rngDevice update with other device
updates in a single list. Optional<T> class may be used as well for
other fields in VmManagementParametersBase that use the same pattern.
Watchdog is one of them.

Change-Id: If9aa41b0cacd77120b002123d8e196e215e20fdc
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>