webadmin: prevent class cast exception on import vm
This patch is a workaround for a problem with StorageDisksListModel that
cause class-cast exception: in ImportVmFromExportDomainModel we set the
items in StorageDisksListModel to StorageDomains. However, in
StorageDisksListModel#updateActionAvailability we expect the
getSelectedItems method to return DiskImages.

Probably something went wrong in b13dc92e where StorageDisksListModel
was introduced in ImportVmFromExportDomainModel. This patch is only a
workaround that eliminates the error by clearing the storage-domain
before loading new ones so the previous selected item will be null and
we won't get that exception. Proper solution should be introduced
separately.

Change-Id: I28f9eee614ba5e6eeef9f8aa88337aab9a90a2f9
Signed-off-by: Arik Hadas <ahadas@redhat.com>