restapi: Fix possible version mismatch (#1206068)
GET on the root of the API returns:
  <api>
  .
  .
    <version>...</version>
    <full_version>...</full_version>
  </api>

The first is taken from vdc_options-->VdcVersion.
The second is taken from vdc_options-->ProductRPMVersion.

Sometimes the second value is more accurate and there can be a mismatch.
This patch adds a query to the engine that gets the product version.
The query attempst to parse ProductRPMVersion and use it for both fields.
If parsing fails, VdcVersion is used as fall-back.

Bug-Url: https://bugzilla.redhat.com/1206068
Change-Id: I1bd123c2352db229e52bdbd3f1e2db4ac72018a9
Signed-off-by: Ori Liel <oliel@redhat.com>