webadmin: fix import-vm from vms grid
Class-cast exception occurred when import VM from the VMs grid. This
patch fixes it by changing the code that is executed after triggering
the import operation so that no popup will be displayed when the dialog
is closed. That is ok because we are not used to show such dialogs that
tell the user that there are tasks in the background so we can decide
not to show one for the new import dialog which is triggered from the
VMs-grid.

Change-Id: I4f505975e4301594aadbb16cbda3a613c4390888
Signed-off-by: Arik Hadas <ahadas@redhat.com>