core: CommandBase - avoid releasing lock if there is a callback
Currently command can be marked to keep it's acquired lock after the
execute phase ends. However, if it doesn't have tasks the locks will be
released regardless.

This patch changes that, if it was marked to keep the locks after the
execute phase and there is a callback, it's the callback responsibllity
to run the command endMethod() once polling has finished to release the
acquired locks.

Change-Id: If8600ee529f68a1b3e2881f4e33f44fec9402beb
Signed-off-by: laravot@redhat.com <laravot@redhat.com>