PendingResourceManager for tracking resources in WaitForLaunch
This patch adds a new scheduler component PendingResourceManager.

This manager is responsible for keeping track of all resources
that are allocated by not-yet running VMs. Each resource type
is represented by its own class type (and values) and can compute
summary information about pending counts/sizes per host.

This should allow us to easily track multiple resources we lack
today - VMs themselves (for affinity or PCI passthrough), VNICs
and others.

The benefit of this patch is that multiple calls to PRM.cleanVm()
do not cause the counters to be decremented multiple times. The
manager tracks all pending resources separately and has indexes
for per VM and per Host access.

The old VDS fields for pending memory and CPU are left in the
entity for cache purposes. Reading them does not guarantee
validity though. Only SchedulingManager's instance of PRM
guaranteed up-to-date and valid values.

The only procedure which is allowed to put those to DB is the
VdsManager via PendingResourceManager. VdsManager has a new call
updatePendingData to receive the current pending values from PRM.

Change-Id: Ie741b8e244acb64e470af83252d90ec134ba7f8e
Signed-off-by: Martin Sivak <msivak@redhat.com>