engine, webadmin: Check volume emptyness for geo-rep
This patch adds VDS command to check emptyness
of gluster volume.
And also adds a predicate in the GlusterGeoRepUtil
to check the same while filtering the volumes
list to get list of volumes eligible for geo-rep
session create.
This patch also adds a UIMessage entry corresponding
to the new volume emptyness criteria.
Change-Id: Ia86df17975069d4fe7c1740bf609597fde7b08ac
Signed-off-by: Anmol Babu <anbabu@redhat.com>