engine : Sync volume details on creation
To create a new geo-rep session user it is
required for the volumes to have their capacity
information. However for a case where the user
has just created the volume and wants this volume
to participate in a geo-rep session, since the
volume earlier didn't carry the capacity info,
such a volume wouldn't feature in the eligible
volumes list in the create geo-rep session dialog.
Hence, this patch tries to sync the capacity info
of a volume on its successful creation.

Change-Id: Ic4d78ab70c0c28f107b589c1b6ff14848b184928
Signed-off-by: Anmol Babu <anbabu@redhat.com>