core: always send numa node for vm
in case user didn't specify numa config for a vm,
send default numa configuration for the vm,
using the host's first numa node.

this is done in clusters that support numa and for hosts that has a numa
node.

this is needed for up coming hot-plug memory for vms,
that require at least one numa node in the guest.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1224886
Change-Id: I9c7b0938bb0ab45883d637b594de61cd8695963a
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>