core: add MemoryStorageHandler the ability to sort
This patch adds the MemoryStorageHandler mechanism the ability to sort
the list of storage domains according to given comparators logic. For
example, sort by the available disk size of a storage domain.
This way, if we have more than one valid candidate, we can sort the list
and get the most suitable domain out of it.
When there is more than one comparator, a nested sort is performed.

Change-Id: I00ad78fca97527a7dbf426a28ced7314a7bd921c
Bug-Url: https://bugzilla.redhat.com/1186230
Signed-off-by: Idan Shaby <ishaby@redhat.com>