userportal,webadmin: all disks were by default bootable for instance images
The problem was that the new disk dialog opened from the new VM dialog was
always "bootable: true".

Fixed by 2 changes:
- overridden the AbstractDiskModel.updateBootableDiskAvailable() in instance
images models. The difference is that in the default implementation it takes
the disks from server while in case of instance images from the dialg (not yet
saved changes)

- if created the disk image from the VM dialog (e.g. instance image) it is set
"active" since it has to be considered to be active (e.g. no other disks has
tbe bootable)

Change-Id: Ifb699b2c1d8c2124b69d8032eefd8ba8e185844f
Bug-Url: https://bugzilla.redhat.com/1213257
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>