restapi: Don't display PM type when PM disabled (#1191478)
Currenly Power-Management type is disaplyed even when
Power-Managemnent is disabled for the host. This patch
fixed that.

This patch also declares this field as deprecated, since
speaking of the 'type' of Power-Management doesn't make
sense since intorduction of fence-agents. Type should be
associated with a specific agent, not the entire PM.
It was not deleted from PM to maintain backwards
compatibility.

Bug-Url: http://bugzilla.redhat.com/1191478
Change-Id: I0630ebde2b342964f44c3fad51de1ca5032b5c23
Signed-off-by: Ori Liel <oliel@redhat.com>