webadmin: drop import-as-clone popups
These popup dialogs were complicated and not intuitive, so they are
replaced by a different ui mechanism in the import dialog that allows
the user to configure a different name for cloned VMs.

This mechanism also works for VMs that are being imported not as a
clone.

The validation on the widget for the VM name is missing in this patch
and will be added separately (it means that currently the user doesn't
get an accurate error message).

Change-Id: I31b8cee780ce5f95a6f4085255927847b01ff994
Signed-off-by: Arik Hadas <ahadas@redhat.com>