core: ExtendSANStorageDomain storage type check
ExtendSANStorageDomainCommand should only be applied on block storage
domains.
The current canDoAction() is wrong, as it theoritically[1] allows the
command to be run on POSIX, Gluster and LocalFS storage domains.

[1] Note: With the current code base this patch has no functional gain,
as this condition is checked in every flow that can potentially call
this command (both GUI and REST), but it is the correct thing to do, and
may serve to prevent future bugs if this command is ever called from a
different flow.

Change-Id: If25b65ba482a1659d5083f06dd76fa42c68e971e
Bug-Url: https://bugzilla.redhat.com/1213969
Signed-off-by: Allon Mureinik <amureini@redhat.com>