restapi: Don't override potentially existing configuration
Currently the mapping method that populates the configuration of a VM
(the OVF configuration, for example) doesn't take into account that it
may already contain other values, like cloud-init or sysprep
configuration.  As a result the existing configuration is completely
replaced, and a wrong result is returned to the user. This patch changes
that mapping method so that it uses the existing configuration should it
exist.

Change-Id: Ica89aac29e19e959f0a44402f2715f18de984c1f
Bug-Url: https://bugzilla.redhat.com/1218316
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>