core: use native launch as the default console mode
Changes the default spice console mode from auto (e.g. try plugin, if not there
use native client) to use native by default.

This change will not touch the existing installations during update only the
new installations, since this is an option configurable by user so we are not
destroying it.

Change-Id: I4e63a2386d7bd43757c34b7839a6b4e3ca3520b2
Bug-Url: https://bugzilla.redhat.com/1083523
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>