webadmin: fluent UiAction
This patch includes action classes that can assist running vdc action
(and may be queries in the future) in a fluent way.

An action can be dependent on the previous action result or can run
parallely to it.

The mechanism takes care of starting the model progress when running the
first action and stopping in after the last action result was return.

It also takes care of collecting the actions failures and displaying them
in one error popup.

A flow example-
actionA.then(actionB).and(actionC).and(actionD).then(actionE).and(actionF).
onAllExecutionsFinish(actionG);
actionA.runAction();

actionA ->actionB
        ->actionC
        ->actionD->actionE
                 ->actionF
                          ->actionG

Change-Id: I262282d00bbb16a65f36a2463d3ffe8dbf6594c6
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>